Propagate invalidate_corrupt to signature validation
I set invalidate_corrupt to default to True inside SignedCookie
for backwards compatibility reasons (if any client code instantiates
SignedCookie directly). But this will extend the handling of
invalidate_corrupt to a case it did not previously handle. Before this,
if invalidate_corrupt was False, it was still possible to silently
invalidate the session if the signature was invalid.