Allow using pycryptodome, and prefer it over pycrypto
pycryptodome is a maintained fork of pycrypto. They are mostly
API-compatible (and use the same package name, i.e. meant to be
a drop-in replacement), and AFAICS the little bit of API that Beaker
uses is fully compatible.

Therefore, list pycryptodome as another extra dependency. Furthermore,
since pycrypto is dead and being phased out whenever possible, use
pycryptodome for tests instead of it.