Fix docs building and document Redis/MongoDB
The conf.py contained theme settings that broke the build with lates Sphinx and
that are also not relevant to the now used RTD theme.

Mention Redis and MongoDB as available backends since I've missed them myself
when looking at the docs.

Fix markup in one docstring that generated an RST error.