Fix path for CookieSession
CookieSession does not properly apply the `cookie_path` configuration
option, causing the resulting cookie to be set for the default path
instead.

This happens on several occasions: after first initialization, after
calling `delete` and on subsequent initializations when the cookie is
loaded.

Properties that are inherent to the cookie itself (such as domain and
path) are stored as both dict items and object attributes. The dict
items are used to persist the properties in the session state so that
the they can be retrieved on load, while the object attributes are used
internally. Both of these need to be respected when the cookie is
created.