Make md5 calls succeed on systems enforcing FIPS compliance.
The Federal Information Processing Standards (FIPS) do not allow using the md5 algorithm for encryption since it is considered too weak. The hashlib implementation of md5 provided a workaround permitting continued use of the fast md5 algorithm by explicitly noting the use was not for security purposes. The changes here try using that workaround, and fall back to failing on systems enforcing FIPS.